lib: Fix double/int comparisons in a few progress reporting functions
Spotted here: https://gitlab.gnome.org/GNOME/gnome-software/-/merge_requests/1940#note_2073120

+++ lib/gs-plugin-job-refresh-metadata.c
@@ -350,7 +350,7 @@ progress_cb (gpointer user_data)
 	progress += (100.0 / n_portions) * external_appstream_completion;
 #endif
 
-	if (progress != self->last_reported_progress) {
+	if ((guint) progress != self->last_reported_progress) {
 		g_signal_emit (self, signals[SIGNAL_PROGRESS], 0, (guint) progress);
 		self->last_reported_progress = progress;
 	}
